Renamed EndPoint to EndPointInfo. Also moved EndPoint related stuff to its own package in bridge. Also created a serialize package, moved EndPointSerializer to it.
